+++ packages/create-sitecore-jss/src/templates/angular/src/app/lib/layout-service-factory.ts
@@ -18,7 +18,7 @@ export class LayoutServiceFactory {
           apiHost: environment.sitecoreApiHost,
           apiKey: environment.sitecoreApiKey,
           siteName: environment.jssAppName,
-          configurationName: 'default',
+          configurationName: environment.configurationName,
         });
   }
 }