+++ packages/create-sitecore-jss/src/templates/react/scripts/generate-config.js
@@ -16,6 +16,7 @@ const defaultConfig = {
   graphQLEndpointPath: process.env[`${constantCase('reactAppGraphQLEndpointPath')}`],
   defaultLanguage: process.env[`${constantCase('reactAppDefaultLanguage')}`],
   graphQLEndpoint: process.env[`${constantCase('reactAppGraphQLEndpoint')}`],
+  configurationName: process.env[`${constantCase('configurationName')}`],
 };
 
 generateConfig()